refactor(E2eiEnrollment): move crypto impls to own module
