+++ crypto/src/e2e_identity/crypto.rs
@@ -1,40 +1,10 @@
 use super::error::*;
 use crate::{MlsError, prelude::MlsCiphersuite};
-use mls_crypto_provider::{MlsCryptoProvider, PkiKeypair, RustCrypto};
+use mls_crypto_provider::PkiKeypair;
 use openmls_basic_credential::SignatureKeyPair as OpenMlsSignatureKeyPair;
-use openmls_traits::{
-    OpenMlsCryptoProvider,
-    crypto::OpenMlsCrypto,
-    types::{Ciphersuite, SignatureScheme},
-};
+use openmls_traits::types::{Ciphersuite, SignatureScheme};
 use wire_e2e_identity::prelude::JwsAlgorithm;
 use zeroize::Zeroize;
-
-impl super::E2eiEnrollment {
-    pub(super) fn new_sign_key(
-        ciphersuite: MlsCiphersuite,
-        backend: &MlsCryptoProvider,
-    ) -> Result<E2eiSignatureKeypair> {
-        let (sk, _) = backend
-            .crypto()
-            .signature_key_gen(ciphersuite.signature_algorithm())
-            .map_err(MlsError::wrap("performing signature keygen"))?;
-        E2eiSignatureKeypair::try_new(ciphersuite.signature_algorithm(), sk)
-    }
-
-    pub(crate) fn get_sign_key_for_mls(&self) -> Result<Vec<u8>> {
-        let sk = match self.ciphersuite.signature_algorithm() {
-            SignatureScheme::ECDSA_SECP256R1_SHA256 | SignatureScheme::ECDSA_SECP384R1_SHA384 => self.sign_sk.to_vec(),
-            SignatureScheme::ECDSA_SECP521R1_SHA512 => RustCrypto::normalize_p521_secret_key(&self.sign_sk).to_vec(),
-            SignatureScheme::ED25519 => RustCrypto::normalize_ed25519_key(self.sign_sk.as_slice())
-                .map_err(MlsError::wrap("normalizing ed25519 key"))?
-                .to_bytes()
-                .to_vec(),
-            SignatureScheme::ED448 => return Err(Error::NotYetSupported),
-        };
-        Ok(sk)
-    }
-}
 
 impl TryFrom<MlsCiphersuite> for JwsAlgorithm {
     type Error = Error;
