+++ crypto/src/e2e_identity/enrollment/mod.rs
@@ -1,3 +1,5 @@
+mod crypto;
+
 use core_crypto_keystore::CryptoKeystoreMls as _;
 use mls_crypto_provider::MlsCryptoProvider;
 use openmls_traits::{OpenMlsCryptoProvider as _, random::OpenMlsRand as _};
